+++ R/relRisk.R
@@ -20,7 +20,7 @@
 #' 
 #' @examples
 #'\dontrun{
-#' # Analyzing the BIGDAWG::HLA_data dataset #'
+#' # Analyzing the BIGDAWG::HLA_data dataset
 #' library(BIGDAWG)
 #' rr <- relRisk(HLA_data[,1:4])
 #' }